Splitting LED light into subclass so can have multiple lights
